Update dependencies and refactor template placeholder handling
This commit updates the version of the nascore_util dependency in both the PubFrontEnd and vod modules. It also introduces the replacetemplateplaceholders package to handle template placeholder replacements in the ServeStaticFile and FileServerEmbed functions, improving code modularity and maintainability. Additionally, the notAdmin.html template has been enhanced with a more structured layout and improved user experience for non-admin users.